requirements.txt
Flask==2.1.0
gunicorn==20.1.0
line-bot-sdk==2.3.0
APScheduler==3.9.1
psycopg2==2.9.3
pandas==1.4.3